+++ monocle/db/queries.py
@@ -107,6 +107,7 @@ def generate_events_filter(params, qfilter):
 def generate_changes_filter(params, qfilter):
     state = params.get('state')
     tests_included = params.get('tests_included')
+    self_merged = params.get('self_merged')
     has_issue_tracker_links = params.get('has_issue_tracker_links')
     if state:
         qfilter.append({"term": {"state": state}})
@@ -127,6 +128,8 @@ def generate_changes_filter(params, qfilter):
                     }
                 }
             )
+    if self_merged:
+        qfilter.append({"term": {"self_merged": True}})
 
 
 def generate_filter(es, index, repository_fullname, params, ensure_time_range=True):